+++ src/main/java/com/TatianaCarvajal/HotelAPI/model/Reservation.java
@@ -12,7 +12,7 @@ public class Reservation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reservationId;
+    private long reservationId;
 
     private String checkInDate;
     private String checkOutDate;
